BUGFIX: Make devserver workagain
