+++ .gitignore
@@ -7,6 +7,8 @@
 # Application & development dependencies.
 #
 node_modules
+.yarn/cache
+.yarn/install-state.gz
 
 #
 # Compiled assets.
