+++ .testcaferc.json
@@ -1,5 +1,5 @@
 {
   "browsers": "chromium",
   "src": ["Resources/Private/JavaScript/media-module/tests/**/*.ts"],
-  "appCommand": "yarn ts-node Resources/Private/JavaScript/dev-server/server.ts"
+  "appCommand": "yarn ts-node --project Resources/Private/JavaScript/dev-server/tsconfig.json Resources/Private/JavaScript/dev-server/server.ts"
 }
