+++ Resources/Private/JavaScript/dev-server/package.json
@@ -3,9 +3,12 @@
   "version": "1.0.0",
   "license": "GNU GPLv3",
   "dependencies": {
-    "@media-ui/media-module": "*"
+    "@media-ui/media-module": "*",
+    "@media-ui/feature-asset-usage": "*",
+    "@media-ui/feature-concurrent-editing": "*"
   },
   "devDependencies": {
+    "@types/node": "^16.18.3",
     "express": "^4.17.1",
     "nodemon": "^2.0.6"
   },
