+++ Resources/Private/JavaScript/dev-server/src/index.html
@@ -5,6 +5,7 @@
     <meta content="width=device-width, initial-scale=1.0" name="viewport">
     <meta content="ie=edge" http-equiv="X-UA-Compatible">
     <title>Media UI - Dev server</title>
+    <script>var parcelRequire;</script>
 </head>
 
 <body class="neos neos-module neos-controls">
@@ -20,7 +21,9 @@
     document.head.appendChild(style);
 </script>
 
+<script>var exports = {};</script>
+
 <script src="./index.ts"></script>
-<script src="../media-module/src/index.tsx"></script>
+<script src="../../media-module/src/index.tsx"></script>
 </body>
 </html>
