+++ Resources/Private/JavaScript/dev-server/tsconfig.json
@@ -1,8 +1,23 @@
 {
   "extends": "../../../../tsconfig.json",
+  "include": [
+    "src/**/*.ts",
+    "node_modules/@media-ui/feature-asset-usage/src/**/*",
+    "node_modules/@media-ui/feature-asset-variants/src/**/*",
+    "node_modules/@media-ui/feature-concurrent-editing/src/**/*"
+  ],
+  "exclude": [
+    "node_modules"
+  ],
   "compilerOptions": {
     "rootDir": "src",
-    "module": "CommonJS",
-    "jsx": "react"
+    "jsx": "react",
+    "module": "esnext",
+    "esModuleInterop": true,
+    "moduleResolution": "node",
+    "types": ["node"]
+  },
+  "ts-node": {
+    "esm": true
   }
 }
