+++ Resources/Private/JavaScript/media-module/package.json
@@ -67,9 +67,7 @@
     "postcss-modules": "^1.5.0",
     "postinstall-postinstall": "^2.1.0",
     "react-hot-loader": "^4.13.0",
-    "testcafe": "^1.10.0",
-    "ts-node": "^9.1.1",
-    "typescript": "^4.1.3"
+    "typescript": "^4.9.3"
   },
   "alias": {
     "@neos-project/react-ui-components/*": "@neos-project/react-ui-components/lib-esm/*"
